Remove i32 and u32 variants of DataType
Curretly noria uses four integer types, i32, u32, i64 and u64 internally,
this is in contrast to SQL that also has i/u8 and i/u16 bit variants, so
it is not entirely clear what i32 and u32 bring to the table on 64 bit
machines.

Performance: on 64 bit machines all operations are performed on 64bits anyway,
             moreover having fewer types greatly reduces the number of branches
             in our logic which should lead to improved performance.
Correctness: if we can get correct results for i8 and i16 using i32, no reason
             we can't use i64 for i32 (same for unsigned variants).
Serialization: all integer types are serialized as a variable length i128, and
               deserialized to the closest type, which in fact can lead to more
               logic bugs.

Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1836
Tested-by: Buildkite CI

+++ noria-psql/src/resultset.rs
@@ -172,7 +172,7 @@ mod tests {
     #[test]
     fn iterate_resultset() {
         let results = vec![Results::new(
-            vec![vec![DataType::BigInt(10)]],
+            vec![vec![DataType::Int(10)]],
             Arc::new(["col1".to_string()]),
         )];
         let schema = SelectSchema(cl::SelectSchema {
@@ -199,12 +199,12 @@ mod tests {
     fn iterate_resultset_with_multiple_results() {
         let results = vec![
             Results::new(
-                vec![vec![DataType::BigInt(10)]],
+                vec![vec![DataType::Int(10)]],
                 Arc::new(["col1".to_string()]),
             ),
             Results::new(Vec::<Vec<DataType>>::new(), Arc::new(["col1".to_string()])),
             Results::new(
-                vec![vec![DataType::BigInt(11)], vec![DataType::BigInt(12)]],
+                vec![vec![DataType::Int(11)], vec![DataType::Int(12)]],
                 Arc::new(["col1".to_string()]),
             ),
         ];
@@ -280,14 +280,14 @@ mod tests {
         let results = vec![Results::new(
             vec![
                 vec![
-                    DataType::BigInt(10),
-                    DataType::BigInt(99),
+                    DataType::Int(10),
+                    DataType::Int(99),
                     DataType::Text("abcdef".into()),
                     DataType::Int(0),
                 ],
                 vec![
-                    DataType::BigInt(11),
-                    DataType::BigInt(99),
+                    DataType::Int(11),
+                    DataType::Int(99),
                     DataType::Text("ghijkl".into()),
                     DataType::Int(0),
                 ],
